Move apply command stuff to its own package
