+++ bin/test/io/github/alechenninger/monarch/ConfigSerializationTest.groovy
@@ -18,6 +18,7 @@
 
 package io.github.alechenninger.monarch
 
+import io.github.alechenninger.monarch.apply.ApplyChangesOptionsFromSerializableConfig
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
